health\lives handling in healthdamage system

+++ src/engine/systems/HealthDamage.java
@@ -43,7 +43,7 @@ public class HealthDamage implements ISystem {
 	}
 
     public void addComponent(int pid, String componentName) {
-		if(!componentName.equals(Health.KEY) && !componentName.equals(DamageLauncher.KEY)) {
+		if(!componentName.equals(Health.KEY) && !componentName.equals(Damage.KEY)) {
 			return;
 		}
 		
@@ -108,8 +108,16 @@ public class HealthDamage implements ISystem {
 				}
 
 				if(h.getHealth() <= 0) {
-					em.removeEntity(key);
-					System.out.println("removing");
+					Player p = (Player) em.getComponent(h.getParentID(), Player.KEY);
+					if(p!=null) {
+					p.setLives(p.getLives()-1);	
+					h.resetHealth();					
+					}
+					else {
+						em.removeEntity(key);
+						System.out.println("removing");
+					}
+					
 				}
 
 			}
